+++ website/src/views/View.ts
@@ -1,12 +1,11 @@
 import type { DateRangeOption } from '@genspectrum/dashboard-components/util';
 
-import type { OrganismConstants, SequencingEffortsConstants, SingleVariantConstants } from './OrganismConstants.ts';
-import { type PageStateHandler, SequencingEffortsStateHandler, SingleVariantStateHandler } from './PageStateHandler.ts';
-import { sequencingEffortsViewConstants, singleVariantViewConstants, type ViewConstants } from './ViewConstants';
+import type { OrganismConstants } from './OrganismConstants.ts';
+import { type PageStateHandler } from './PageStateHandler.ts';
+import { type ViewConstants } from './ViewConstants';
 import type { LapisLineageQuery, LapisLocation, LapisMutationQuery } from './helpers.ts';
 import type { LineageFilterConfig } from '../components/pageStateSelectors/VariantSelector.tsx';
-import { type BreadcrumbElement, defaultBreadcrumbs } from '../layouts/Breadcrumbs.tsx';
-import { organismConfig } from '../types/Organism.ts';
+import { type BreadcrumbElement } from '../layouts/Breadcrumbs.tsx';
 
 export type BaselineFilter = {
     location: LapisLocation;
@@ -57,81 +56,6 @@ export type View<
 export const defaultTablePageSize = 200;
 
 export const pathoplexusGroupNameField = 'groupName';
-
-export abstract class BaseView<
-    PageState extends object,
-    Constants extends OrganismConstants,
-    StateHandler extends PageStateHandler<PageState>,
-> implements View<PageState, Constants, StateHandler>
-{
-    public readonly pathname;
-    public readonly viewTitle;
-    public readonly viewBreadcrumbEntries;
-
-    protected constructor(
-        public readonly organismConstants: Constants,
-        public readonly pageStateHandler: StateHandler,
-        public readonly viewConstants: ViewConstants,
-    ) {
-        this.pathname = `/${organismConfig[this.organismConstants.organism].pathFragment}/${this.viewConstants.pathFragment}`;
-        this.viewTitle = `${this.viewConstants.label} | ${organismConfig[this.organismConstants.organism].label} | GenSpectrum`;
-        this.viewBreadcrumbEntries = [
-            ...defaultBreadcrumbs,
-            { name: organismConfig[this.organismConstants.organism].label },
-            { name: this.viewConstants.label, href: this.pathname },
-        ];
-    }
-}
-
-export class GenericSingleVariantView<Constants extends SingleVariantConstants> extends BaseView<
-    BaselineAndVariantData,
-    Constants,
-    SingleVariantStateHandler
-> {
-    constructor(constants: Constants) {
-        super(
-            constants,
-            new SingleVariantStateHandler(
-                constants,
-                {
-                    baselineFilter: {
-                        location: {},
-                        dateRange: constants.defaultDateRange,
-                    },
-                    variantFilter: {
-                        mutations: {},
-                        lineages: {},
-                    },
-                },
-                organismConfig[constants.organism].pathFragment,
-            ),
-            singleVariantViewConstants,
-        );
-    }
-}
-
-export class GenericSequencingEffortsView<Constants extends SequencingEffortsConstants> extends BaseView<
-    BaselineData,
-    Constants,
-    SequencingEffortsStateHandler
-> {
-    constructor(constants: Constants) {
-        super(
-            constants,
-            new SequencingEffortsStateHandler(
-                constants,
-                {
-                    baselineFilter: {
-                        location: {},
-                        dateRange: constants.defaultDateRange,
-                    },
-                },
-                organismConfig[constants.organism].pathFragment,
-            ),
-            sequencingEffortsViewConstants,
-        );
-    }
-}
 
 export function getLineageFilterFields(lineageFilters: LineageFilterConfig[]) {
     return lineageFilters.map((filter) => filter.lapisField);
